Update rails_renderer.rb
The content-type of the response for a JSONP request should be "text/javascript" (e.g., see: http://stackoverflow.com/a/111319). Otherwise, if the browser has enabled strict MIME type checking, it will NOT execute the response and thus the request will fail (e.g., see: http://webmasters.stackexchange.com/questions/50006/chrome-refused-to-execute-this-javascript-file).